# openapi/flowai-api.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "FlowAI Business Platform API",
    "description": "ERP workflow portfolio API prepared for Power Platform Custom Connector and Copilot Studio Agent integration.",
    "version": "v1"
  },
  "paths": {
    "/api/agent/contracts/pending-approval": {
      "get": {
        "tags": [
          "Agent"
        ],
        "summary": "Gets contracts that are waiting for approval or not yet approved.",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Contract"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Contract"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Contract"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKey": [ ]
          }
        ]
      }
    },
    "/api/agent/workorders/today": {
      "get": {
        "tags": [
          "Agent"
        ],
        "summary": "Gets work orders currently in progress today.",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WorkOrder"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WorkOrder"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WorkOrder"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKey": [ ]
          }
        ]
      }
    },
    "/api/agent/workorders/delayed": {
      "get": {
        "tags": [
          "Agent"
        ],
        "summary": "Gets delayed work order candidates using the current temporary delay rule.",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WorkOrder"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WorkOrder"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WorkOrder"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKey": [ ]
          }
        ]
      }
    },
    "/api/agent/settlements/on-hold": {
      "get": {
        "tags": [
          "Agent"
        ],
        "summary": "Gets settlements currently on hold.",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Settlement"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Settlement"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Settlement"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKey": [ ]
          }
        ]
      }
    },
    "/api/agent/customers/{customerName}/status-summary": {
      "get": {
        "tags": [
          "Agent"
        ],
        "summary": "Gets workflow status summary for a customer.",
        "parameters": [
          {
            "name": "customerName",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerStatusSummaryResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerStatusSummaryResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerStatusSummaryResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKey": [ ]
          }
        ]
      }
    },
    "/api/agent/query": {
      "post": {
        "tags": [
          "Agent"
        ],
        "summary": "Returns a RuleBased or Ollama-generated natural-language answer from current in-memory workflow data.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentQueryRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentQueryRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AgentQueryRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AgentQueryResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentQueryResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentQueryResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKey": [ ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AgentQueryFactsResponse": {
        "type": "object",
        "properties": {
          "pendingApprovalContracts": {
            "type": "integer",
            "format": "int32"
          },
          "todayWorkOrders": {
            "type": "integer",
            "format": "int32"
          },
          "delayedWorkOrders": {
            "type": "integer",
            "format": "int32"
          },
          "onHoldSettlements": {
            "type": "integer",
            "format": "int32"
          },
          "totalSettlements": {
            "type": "integer",
            "format": "int32"
          },
          "requestedSettlements": {
            "type": "integer",
            "format": "int32"
          },
          "reviewingSettlements": {
            "type": "integer",
            "format": "int32"
          },
          "approvedSettlements": {
            "type": "integer",
            "format": "int32"
          },
          "rejectedSettlements": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "AgentQueryRequest": {
        "type": "object",
        "properties": {
          "question": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "AgentQueryResponse": {
        "type": "object",
        "properties": {
          "question": {
            "type": "string",
            "nullable": true
          },
          "answer": {
            "type": "string",
            "nullable": true
          },
          "source": {
            "type": "string",
            "nullable": true
          },
          "generatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "facts": {
            "$ref": "#/components/schemas/AgentQueryFactsResponse"
          }
        },
        "additionalProperties": false
      },
...
      "BusinessStatus": {
        "enum": [
          10,
          20,
          30,
          31,
          32,
          33,
          40,
          50,
          60,
          70,
          80,
          90,
          100,
          101,
          110,
          120,
          130
        ],
        "type": "integer",
        "format": "int32"
      },
      "Contract": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "contractNo": {
            "type": "string",
            "nullable": true
          },
          "customerName": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "amount": {
            "type": "number",
            "format": "double"
          },
          "status": {
            "$ref": "#/components/schemas/BusinessStatus"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "approvedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "CustomerStatusSummaryResponse": {
        "type": "object",
        "properties": {
          "customerName": {
            "type": "string",
            "nullable": true
          },
          "contractCount": {
            "type": "integer",
            "format": "int32"
          },
          "approvedContractCount": {
            "type": "integer",
            "format": "int32"
          },
          "workOrderCount": {
            "type": "integer",
            "format": "int32"
          },
          "workInProgressCount": {
            "type": "integer",
            "format": "int32"
          },
          "workCompletedCount": {
            "type": "integer",
            "format": "int32"
          },
          "settlementCount": {
            "type": "integer",
            "format": "int32"
          },
          "settlementApprovedCount": {
            "type": "integer",
            "format": "int32"
          },
          "settlementHeldCount": {
            "type": "integer",
            "format": "int32"
          },
          "summary": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Settlement": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "workOrderId": {
            "type": "integer",
            "format": "int32"
          },
          "settlementNo": {
            "type": "string",
            "nullable": true
          },
          "amount": {
            "type": "number",
            "format": "double"
          },
          "status": {
            "$ref": "#/components/schemas/BusinessStatus"
          },
          "holdReason": {
            "type": "string",
            "nullable": true
          },
          "requestedAt": {
            "type": "string",
            "format": "date-time"
          },
          "approvedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "WorkOrder": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "contractId": {
            "type": "integer",
            "format": "int32"
          },
          "workNo": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/BusinessStatus"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "startedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "completedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "ApiKey": {
        "type": "apiKey",
        "description": "Enter the FlowAI API key. Example: local-dev-key",
        "name": "X-FlowAI-Api-Key",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "ApiKey": [ ]
    }
  ],
...
}